+++ sources/hocon/src/main/java/io/smallrye/config/source/hocon/HoconConfigSourceProvider.java
@@ -1,4 +1,4 @@
-package io.smallrye.configsource;
+package io.smallrye.config.source.hocon;
 
 import java.util.*;
