+++ sources/hocon/src/main/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSourceProvider
@@ -1,1 +1,1 @@
-io.smallrye.configsource.HoconConfigSourceProvider
+io.smallrye.config.source.hocon.HoconConfigSourceProvider
